now better handling case-insensitive file names as needed for windows
The algorithm or instance name can contain uppercase and lowercase characters.
The file names for log files are based on the algorithm and instance name.
The Windows file systems may be case-insensitive, meaning that they treat uppercase and lowercase characters the same, i.e., the file name "ABC.txt" is the same as "abc.txt".
The canonical file names that Python provides may then always be lowercase, as it does not matter for the file system.
However, our system demands that the filenames of log files are consistent with the algorithm and instance names and that their cases match.
Otherwise, they throw an error.
This mechanism was part of our fight against any form of corrupt data.
However, we now permit that there are differences in letter case in the file names compared to the algorithm and instance names.
Then our log-based evaluation also works on case-insensitive file systems such as FAT and others used, e.g., under Windows.