now no longer using numba for op1_swap_2_or_flip
It seems that there might be a memory leak that affects the use of numpy random Generators with numba.
I have noted it as https://github.com/numba/numba/issues/9103
This very likely also affects other moptipy code.
For now, we only exclude numba where it hurts in the present experiments.
Once I understand what is actually going on, this can be fixed.